+++ packages/media/media-picker/src/domain/uploadParams.ts
@@ -2,5 +2,4 @@ import { UploadParams } from './config';
 
 export const defaultUploadParams: UploadParams = {
   collection: '',
-  copyFileToRecents: true,
 };