+++ packages/media/media-picker/src/popup/components/app.tsx
@@ -52,7 +52,8 @@ import { MediaPickerPopupWrapper, SidebarWrapper, ViewWrapper } from './styled';
 export interface AppStateProps {
   readonly selectedServiceName: ServiceName;
   readonly isVisible: boolean;
-  readonly context: Context;
+  readonly tenantContext: Context;
+  readonly userContext: Context;
   readonly config?: Partial<PopupConfig>;
 }
 
@@ -91,7 +92,6 @@ export class App extends Component<AppProps, AppState> {
   private readonly mpBrowser: MpBrowser;
   private readonly mpDropzone: MpDropzone;
   private readonly mpBinary: MpBinary;
-  private readonly userContext: Context;
 
   constructor(props: AppProps) {
     super(props);
@@ -103,31 +103,27 @@ export class App extends Component<AppProps, AppState> {
       onUploadProcessing,
       onUploadEnd,
       onUploadError,
-      context,
+      tenantContext,
+      userContext,
       tenantUploadParams,
     } = props;
-    const { userAuthProvider } = context.config;
 
-    if (!userAuthProvider) {
-      throw new Error('userAuthProvider must be provided in the context');
-    }
     this.state = {
       isDropzoneActive: false,
     };
 
-    const uploadParams: UploadParams = {
-      collection: RECENTS_COLLECTION,
-      copyFileToRecents: false,
-    };
-
-    // We need to create a new context since Cards in recents view need user auth
-    this.userContext = ContextFactory.create({
-      authProvider: userAuthProvider,
+    // Context that has both auth providers defined explicitly using to provided contexts.
+    // Each of the local components using this context will upload first to user's recents
+    // and then copy to a tenant's collection.
+    const context = ContextFactory.create({
+      authProvider: tenantContext.config.authProvider,
+      userAuthProvider: userContext.config.authProvider,
+      cacheSize: tenantContext.config.cacheSize,
     });
 
     this.mpBrowser = MediaPicker('browser', context, {
-      uploadParams,
-      tenantUploadParams,
+      uploadParams: tenantUploadParams,
+      shouldCopyFileToRecents: false,
       multiple: true,
     });
     this.mpBrowser.on('uploads-start', onUploadsStart);
@@ -138,8 +134,8 @@ export class App extends Component<AppProps, AppState> {
     this.mpBrowser.on('upload-error', onUploadError);
 
     this.mpDropzone = MediaPicker('dropzone', context, {
-      uploadParams,
-      tenantUploadParams,
+      uploadParams: tenantUploadParams,
+      shouldCopyFileToRecents: false,
       headless: true,
     });
     this.mpDropzone.on('drag-enter', () => this.setDropzoneActive(true));
@@ -152,8 +148,8 @@ export class App extends Component<AppProps, AppState> {
     this.mpDropzone.on('upload-error', onUploadError);
 
     this.mpBinary = MediaPicker('binary', context, {
-      uploadParams,
-      tenantUploadParams,
+      uploadParams: tenantUploadParams,
+      shouldCopyFileToRecents: false,
     });
     this.mpBinary.on('uploads-start', onUploadsStart);
     this.mpBinary.on('upload-preview-update', onUploadPreviewUpdate);
@@ -222,10 +218,12 @@ export class App extends Component<AppProps, AppState> {
 
   private renderCurrentView(selectedServiceName: ServiceName): JSX.Element {
     if (selectedServiceName === 'upload') {
+      // We need to create a new context since Cards in recents view need user auth
+      const { userContext } = this.props;
       return (
         <UploadView
           mpBrowser={this.mpBrowser}
-          context={this.userContext}
+          context={userContext}
           recentsCollection={RECENTS_COLLECTION}
         />
       );
@@ -243,11 +241,17 @@ export class App extends Component<AppProps, AppState> {
   };
 }
 
-const mapStateToProps = ({ view, context, config }: State): AppStateProps => ({
+const mapStateToProps = ({
+  view,
+  tenantContext,
+  userContext,
+  config,
+}: State): AppStateProps => ({
   selectedServiceName: view.service.name,
   isVisible: view.isVisible,
   config,
-  context,
+  tenantContext,
+  userContext,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch<State>): AppDispatchProps => ({