+++ packages/media/smart-card/src/Client/__tests__/unit/index.test.ts
@@ -23,6 +23,7 @@ function createClient(options?: ClientOptions) {
 
 function resolved() {
   fetchMock.mock({
+    name: 'resolved',
     matcher: `begin:${RESOLVE_URL}`,
     response: {
       status: 200,
@@ -45,6 +46,7 @@ function resolved() {
 
 function errored() {
   fetchMock.mock({
+    name: 'errored',
     matcher: `begin:${RESOLVE_URL}`,
     response: {
       status: 500,
@@ -55,6 +57,7 @@ function errored() {
 
 function notfound() {
   fetchMock.mock({
+    name: 'notfound',
     matcher: `begin:${RESOLVE_URL}`,
     response: {
       status: 200,
@@ -200,7 +203,6 @@ describe('Client', () => {
   it('should reload when reload is called for the same provider', done => {
     expect.assertions(4);
     resolved();
-    resolved();
     let count = 0;
     const client = createClient();
     const subscription = client.get(OBJECT_URL).subscribe(({ status }) => {
@@ -239,7 +241,6 @@ describe('Client', () => {
 
   it('should not reload when reload is called for a different provider', done => {
     expect.assertions(2);
-    resolved();
     resolved();
     let count = 0;
     const client = createClient();