+++ packages/search/global-search/src/components/jira/JiraSearchResultsMapper.ts
@@ -1,31 +1,51 @@
-import { ResultsGroup, JiraResultsMap } from '../../model/Result';
+import {
+  ResultsGroup,
+  JiraResultsMap,
+  GenericResultMap,
+  Result,
+} from '../../model/Result';
 import { take } from '../SearchResultsUtil';
 
 const MAX_ISSUES = 8;
 const MAX_PROJECTS = 2;
 const MAX_BOARDS = 2;
 const MAX_FILTERS = 2;
+const MAX_PEOPLE = 3;
 
-const DEFAULT_JIRA_RESULTS_MAP = {
-  issues: [],
+const DEFAULT_JIRA_RESULTS_MAP: GenericResultMap = {
+  issues: [] as Result[],
   boards: [],
   filters: [],
   projects: [],
+  people: [],
+};
+
+export const sliceResults = (resultsMap: JiraResultsMap) => {
+  const { issues, boards, filters, projects, people } = resultsMap
+    ? resultsMap
+    : DEFAULT_JIRA_RESULTS_MAP;
+
+  const [issuesToDisplay, peopleToDisplay, ...pseudoContainers] = [
+    { items: issues, count: MAX_ISSUES },
+    { items: people, count: MAX_PEOPLE },
+    { items: boards, count: MAX_BOARDS },
+    { items: filters, count: MAX_FILTERS },
+    { items: projects, count: MAX_PROJECTS },
+  ].map(({ items, count }) => take(items, count));
+
+  return {
+    issuesToDisplay,
+    pseudoContainers,
+    peopleToDisplay,
+  };
 };
 
 export const mapRecentResultsToUIGroups = (
   recentlyViewedObjects: JiraResultsMap,
 ): ResultsGroup[] => {
-  const { issues, boards, filters, projects } = recentlyViewedObjects
-    ? recentlyViewedObjects
-    : DEFAULT_JIRA_RESULTS_MAP;
-
-  const [issuesToDisplay, ...others] = [
-    { items: issues, count: MAX_ISSUES },
-    { items: boards, count: MAX_BOARDS },
-    { items: filters, count: MAX_FILTERS },
-    { items: projects, count: MAX_PROJECTS },
-  ].map(({ items, count }) => take(items, count));
+  const { issuesToDisplay, peopleToDisplay, pseudoContainers } = sliceResults(
+    recentlyViewedObjects,
+  );
 
   return [
     {
@@ -34,9 +54,14 @@ export const mapRecentResultsToUIGroups = (
       titleI18nId: 'global-search.jira.recent-issues-heading',
     },
     {
-      items: others.reduce((acc, arr) => [...acc, ...arr]),
+      items: pseudoContainers.reduce((acc, arr) => [...acc, ...arr]),
       key: 'containers',
       titleI18nId: 'global-search.jira.recent-containers',
+    },
+    {
+      items: peopleToDisplay,
+      key: 'people',
+      titleI18nId: 'global-search.jira.recent-people-heading',
     },
   ];
 };
@@ -44,5 +69,24 @@ export const mapRecentResultsToUIGroups = (
 export const mapSearchResultsToUIGroups = (
   searchResultsObjects: JiraResultsMap,
 ): ResultsGroup[] => {
-  return [];
+  const { issuesToDisplay, peopleToDisplay, pseudoContainers } = sliceResults(
+    searchResultsObjects,
+  );
+  return [
+    {
+      items: issuesToDisplay,
+      key: 'issues',
+      titleI18nId: 'global-search.jira.seach-result-issues-heading',
+    },
+    {
+      items: pseudoContainers.reduce((acc, arr) => [...acc, ...arr]),
+      key: 'containers',
+      titleI18nId: 'global-search.jira.seach-result-containers-heading',
+    },
+    {
+      items: peopleToDisplay,
+      key: 'people',
+      titleI18nId: 'global-search.jira.seach-result-people-heading',
+    },
+  ];
 };