+++ packages/elements/emoji/CHANGELOG.md
@@ -1,5 +1,8 @@
 # @atlaskit/emoji
 
+## 40.0.1
+- [patch] FS-2941 Stop using Request object and upgrade fetch-mock [dff332a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/dff332a)
+
 ## 40.0.0
 - [major] Updated dependencies [7545979](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/7545979)
   - @atlaskit/util-data-test@10.0.8